Catalog handles omitted empty lists
If we try creating a catalog entry with an array attribute value giving:
```json
{ "array_value": null }
```

The response from the API will omit the attribute. Similarly, this
payload:
```json
{ "array_value": [] }
```

Will also result in an empty attribute response. The problem is that if
your terraform code had the array_value attribute set to either null or
empty list, terraform will remember the distinction, and when the API
returns a list of attribute values where there is no value, terraform
will perceive this as a state difference and die with the following
error:

```
Error: Provider produced inconsistent result after apply

When applying changes to incident_catalog_entries.Service, provider
"provider[\"registry.terraform.io/incident-io/incident\"]" produced an
unexpected new value:
.entries["attr"].attribute_values["01H7Y6V5P9DP8HMCF2FN48SM6M"].array_value:
was null, but now cty.ListValEmpty(cty.String).
```

To handle this, we have changed how we build the terraform state from
the API response so that when we think it's possible that the API
omitted the attribute, we check our state and use the value in the state
to ensure terraform is happy.