Handle validation errors for 204 endpoints
When the successful response to a client request is 204, it is possible
that we'll parse an error response into successful (empty) result and
not detect that it is an error.

This change adds the validation error handler we've used in other tools
such as the catalog-importer so that we detect errors properly. This was
a report from a customer who saw that Terraform trying to delete a
catalog type resulted in a validation error from our servers but the
catalog type was never deleted (because the endpoint had 422'd).