Make alert source template attributes a set, not a list
This means plan/apply doesn't produce a diff when the order in which you
declare attributes changes, which doesn't matter!

I have tested this by:
- creating a test project
- importing an existing alert source on v4.3.1 (published)
- `terraform plan`-ing with this updated version

It seems like we don't need to write any schema-updating code here,
because set and list are represented in the same way in the state file
(as a JSON list), so this only matters at plan / apply time.