Correctly handle empty slices of team_ids  (#199)
Fixes some bugs:

- One: fix #198. The issue here was that terraform does not return "nil"
if you're internally using a slice of strings. You need to use their
list or set type for this. Because we weren't using types.Set in our
internal model, when doing the connection between the terraform
framework and our API object, it wasn't getting marshalled properly, so
we weren't propagating the empty slice and then validating it out at the
API level (which we should).
* By using the types.Set object, we can make use of the `unknown` and
`nil` properties.
- We were setting managed_attributes in catalog_entries to nil even if
it was empty, which is incorrect. If you set `managed_attributes` to
empty, it means you want it all to be managed in the dashboard.
- We should not be using `Computed` for team_ids, as these should be
explicitly set. We should not let Terraform fallback to what is in the
API response. This also means we cannot have a default value.
- Because of the above, it means we need to make sure we allow schedule
resources to set a nil team_ids when updating, which means the
`buildModel` function had to be updated.

Added unit tests, and also manually ran some terraform plans/applys. 
```
- brand new EP, nil team_ids (errors)  ✅
- brand new EP, empty team_ids (succeeds) ✅
- imports existing EP with no teams, nil team_ids (errors)  ✅
- imports existing EPs with no teams, empty team_ids (succeeds)  ✅
- imports existing EP with existing teams, do not set anything (errors if state is not reconciled) ✅
- import existing schedule that has teams, do not set anything (errors if state is not reconciled) ✅
```