Update catalog_type categories to be computed
With the V3 API we always return an empty array for catalog type
categories. This can cause problem if it's empty as we weren't creating
the stuct from the server response.

This caused a `.categories: was cty.ListValEmpty(cty.String), but now
null.` error. This swaps us out to use a `types.List` type which can
much better handle missing attributes that are computed.