Use ListNestedAttributes for engine conditions
Engine conditions are always stored in the same order that they are sent
to the server, so this should be safe to use list types for them. This
might cause some churn in state files on the first run when the server
is updated to the code order, but after that it should be stable.